Adjust styles for plot. Add documentation for some files. Add qt libs for plotLib. Increase quality of image saved.
